fix: satisfy pkg-file-count limits for session context PR
Move session workflow and executor tests into existing consolidated test
files/subpackages instead of adding new top-level package files.
